requirements.txt
astroid==1.6.5
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0.6
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
gunicorn==19.8.1
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.14.5
openapi-codec==1.3.2
psycopg2==2.7.4
pylint==1.9.2
pytz==2018.4
requests==2.19.1
simplejson==3.15.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
whitenoise==3.3.1
wrapt==1.10.11
